Don’t allow graph downloads on any IE
IE seems very unreliable when it comes to rendering SVG into canvases. For now, disable the Download button on all versions of IE. Uses UA sniffing to do this. Also move BrowserFeat from mixins to its own module.
