+++ src/encoded/static/components/mixins.js
@@ -636,40 +636,3 @@ module.exports.HistoryAndTriggers = {
     }
 
 };
-
-
-// Handle browser capabilities, a la Modernizr. Can *only* be called from
-// mounted components (componentDidMount method would be a good method to
-// use this from), because actual DOM is needed.
-module.exports.BrowserFeat = {
-    feat: {},
-
-    // Return object with browser capabilities; return from cache if available
-    getBrowserCaps: function (feat) {
-        if (Object.keys(this.feat).length === 0) {
-            this.feat.svg = document.implementation.hasFeature('http://www.w3.org/TR/SVG11/feature#Image', '1.1');
-        }
-        return feat ? this.feat[feat] : this.feat;
-    },
-
-    setHtmlFeatClass: function() {
-        var htmlclass = [];
-
-        this.getBrowserCaps();
-
-        // For each set feature, add to the <html> element's class
-        var keys = Object.keys(this.feat);
-        var i = keys.length;
-        while (i--) {
-            if (this.feat[keys[i]]) {
-                htmlclass.push(keys[i]);
-            } else {
-                htmlclass.push('no-' + keys[i]);
-            }
-        }
-
-        // Now write the classes to the <html> DOM element
-        document.documentElement.className = htmlclass.join(' ');
-    }
-};
-
